removed aoSimLib, now use only AOTools
