--- soapy/aoSimLib.py
@@ -1,770 +1,0 @@
-#Copyright Durham University and Andrew Reeves
-#2014
-
-# This file is part of pyAOS.
-
-#     pyAOS is free software: you can redistribute it and/or modify
-#     it under the terms of the GNU General Public License as published by
-#     the Free Software Foundation, either version 3 of the License, or
-#     (at your option) any later version.
-
-#     pyAOS is distributed in the hope that it will be useful,
-#     but WITHOUT ANY WARRANTY; without even the implied warranty of
-#     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#     GNU General Public License for more details.
-
-#     You should have received a copy of the GNU General Public License
-#     along with pyAOS.  If not, see <http://www.gnu.org/licenses/>.
-'''
-A module containing useful functions used throughout pyAOS
-
-:Author:
-    Andrew Reeves
-'''
-
-import numpy
-
-from scipy.interpolate import interp2d,RectBivariateSpline
-#a lookup dict for interp2d order (expressed as 'kind')
-INTERP_KIND = {1: 'linear', 3:'cubic', 5:'quintic'}
-
-from . import AOFFT
-
-#xrange now just "range" in python3.
-#Following code means fastest implementation used in 2 and 3
-try:
-    xrange
-except NameError:
-    xrange = range
-
-def convolve(
-        img1, img2, mode="pyfftw", fftw_FLAGS=("FFTW_MEASURE",),
-        threads=0):
-    '''
-    Convolves two, 2-dimensional arrays
-
-    Uses the AOFFT library to do fast convolution of 2, 2-dimensional numpy ndarrays. The FFT mode, and some parameters can be set in the arguments.
-
-    Parameters:
-        img1 (ndarray): 1st array to be convolved
-        img2 (ndarray): 2nd array to be convolved
-        mode (string, optional): The fft mode used, defaults to fftw
-        fftw_FLAGS (tuple, optional): flags for fftw, defaults to ("FFTW_MEASURE",)
-        threads (int, optional): Number of threads used if mode is fftw
-
-    Returns:
-        ndarray : The convolved 2-dimensional array
-
-    '''
-    #Check arrays are same size
-    if img1.shape!=img2.shape:
-        raise ValueError("Arrays must have same dimensions")
-
-    #Initialise FFT objects
-    fFFT = AOFFT.FFT(img1.shape, axes=(0,1), mode=mode, dtype="complex64",
-                direction="FORWARD", fftw_FLAGS=fftw_FLAGS, THREADS=threads)
-    iFFT = AOFFT.FFT(img1.shape, axes=(0,1), mode=mode, dtype="complex64",
-                  direction="BACKWARD", fftw_FLAGS=fftw_FLAGS, THREADS=threads)
-    #backward FFT arrays
-    iFFT.inputData[:] = img1
-    iImg1 = iFFT().copy()
-    iFFT.inputData[:] = img2
-    iImg2 = iFFT()
-
-    #Do convolution
-    iImg1 *= iImg2
-
-    #do forward FFT
-    fFFT.inputData[:] = iImg1
-    return numpy.fft.fftshift(fFFT().real)
-
-
-def circle(radius, size, centre_offset=(0,0)):
-    """
-    Create a 2-dimensional array equal to 1 in a circle and 0 outside
-
-    Parameters:
-        radius (float): The radius in pixels of the circle
-        size (int): The size of the the array for the circle
-        centre_offset (tuple): The coords of the centre of the circle
-
-    Returns:
-        ndarray : The circle array
-    """
-    size = int(numpy.round(size))
-
-    coords = numpy.linspace(-size/2.,size/2.,size)
-    x,y = numpy.meshgrid(coords,coords)
-    x-=centre_offset[0]
-    y-=centre_offset[1]
-
-    radius+=0.5
-
-    mask = x*x + y*y <= radius*radius
-
-    C = numpy.zeros((size, size))
-    C[mask] = 1
-    return C
-
-def gaussian2d(size, width, amplitude=1., cent=None):
-    '''
-    Generates 2D gaussian distribution
-
-
-    Args:
-        size (tuple, float): Dimensions of Array to place gaussian
-        width (tuple, float): Width of distribution.
-                                Accepts tuple for x and y values.
-        amplitude (float): Amplitude of guassian distribution
-        cent (tuple): Centre of distribution on grid.
-    '''
-
-    try:
-        xSize = size[0]
-        ySize = size[1]
-    except (TypeError, IndexError):
-        xSize = ySize = size
-
-    try:
-        xWidth = float(width[0])
-        yWidth = float(width[1])
-    except (TypeError, IndexError):
-        xWidth = yWidth = float(width)
-
-    if not cent:
-        xCent = size[0]/2.
-        yCent = size[1]/2.
-    else:
-        xCent = cent[0]
-        yCent = cent[1]
-
-    X,Y = numpy.meshgrid(range(0, xSize), range(0, ySize))
-
-    image = amplitude * numpy.exp(
-            -(((xCent-X)/xWidth)**2 + ((yCent-Y)/yWidth)**2)/2)
-
-    return image
-
-
-
-#Interpolation
-###############################################
-
-def zoom(array, newSize, order=3):
-    """
-    A Class to zoom 2-dimensional arrays using interpolation
-
-    Uses the scipy `Interp2d` interpolation routine to zoom into an array. Can cope with real of complex data.
-
-    Parameters:
-        array (ndarray): 2-dimensional array to zoom
-        newSize (tuple): the new size of the required array
-        order (int, optional): Order of interpolation to use. default is 3
-
-    Returns:
-        ndarray : zoom array of new size.
-    """
-
-    if order not in INTERP_KIND:
-       raise ValueError("Order can either be 1, 3, or 5 only")
-
-    try:
-        xSize = newSize[0]
-        ySize = newSize[1]
-
-    except (IndexError, TypeError):
-        xSize = ySize = newSize
-
-    coordsX = numpy.linspace(0, array.shape[0]-1, xSize)
-    coordsY = numpy.linspace(0, array.shape[1]-1, ySize)
-
-    #If array is complex must do 2 interpolations
-    if array.dtype==numpy.complex64 or array.dtype==numpy.complex128:
-
-        realInterpObj = interp2d(   numpy.arange(array.shape[0]),
-                numpy.arange(array.shape[1]), array.real, copy=False,
-                kind=INTERP_KIND[order])
-        imagInterpObj = interp2d(   numpy.arange(array.shape[0]),
-                numpy.arange(array.shape[1]), array.imag, copy=False,
-                kind=INTERP_KIND[order])
-        return (realInterpObj(coordsY,coordsX)
-                            + 1j*imagInterpObj(coordsY,coordsX))
-
-    else:
-
-        interpObj = interp2d(   numpy.arange(array.shape[0]),
-                numpy.arange(array.shape[1]), array, copy=False,
-                kind=INTERP_KIND[order])
-
-        #return numpy.flipud(numpy.rot90(interpObj(coordsY,coordsX)))
-        return interpObj(coordsY,coordsX)
-
-def zoom_rbs(array, newSize, order=3):
-    """
-    A Class to zoom 2-dimensional arrays using RectBivariateSpline interpolation
-
-    Uses the scipy ``RectBivariateSpline`` interpolation routine to zoom into an array. Can cope with real of complex data. May be slower than above ``zoom``, as RBS routine copies data.
-
-    Parameters:
-        array (ndarray): 2-dimensional array to zoom
-        newSize (tuple): the new size of the required array
-        order (int, optional): Order of interpolation to use. default is 3
-
-    Returns:
-        ndarray : zoom array of new size.
-    """
-    try:
-        xSize = newSize[0]
-        ySize = newSize[1]
-
-    except IndexError:
-        xSize = ySize = newSize
-
-    coordsX = numpy.linspace(0, array.shape[0]-1, xSize)
-    coordsY = numpy.linspace(0, array.shape[1]-1, ySize)
-
-    #If array is complex must do 2 interpolations
-    if array.dtype==numpy.complex64 or array.dtype==numpy.complex128:
-        realInterpObj = RectBivariateSpline(
-                numpy.arange(array.shape[0]), numpy.arange(array.shape[1]),
-                array.real, kx=order, ky=order)
-        imagInterpObj = RectBivariateSpline(
-                numpy.arange(array.shape[0]), numpy.arange(array.shape[1]),
-                array.imag, kx=order, ky=order)
-
-        return (realInterpObj(coordsY,coordsX)
-                            + 1j*imagInterpObj(coordsY,coordsX))
-
-    else:
-
-        interpObj = RectBivariateSpline(   numpy.arange(array.shape[0]),
-                numpy.arange(array.shape[1]), array, kx=order, ky=order)
-
-
-        return interpObj(coordsY,coordsX)
-
-
-def interp1d_numpy(array, coords):
-    """
-    A Numpy only inplementation array of 1d interpolation
-
-    Parameters:
-        array (ndarray): The 1d array to be interpolated
-        coords (ndarray): An array of coords to return values
-
-    Returns:
-        ndarray: The interpolated array
-    """
-    intCoords = coords.astype("int")
-    arrayInt = array[intCoords]
-    arrayInt1 = array[(intCoords+1).clip(0,array.shape[0]-1)]
-    grad = arrayInt1 - arrayInt
-
-    rem = coords - intCoords
-
-    interpArray = arrayInt + grad*rem
-
-    return interpArray
-
-
-def interp2d_numpy(array, xCoords, yCoords, interpArray=None):
-    """
-    A Numpy only inplementation array of 2d linear interpolation
-
-    Parameters:
-        array (ndarray): The 2d array to be interpolated
-        xCoords (ndarray): An array of coords to return values
-        yCoords (ndarray): An array of coords to return values
-        interpArray (ndarray, optional): The array to plane the interpolated data
-
-    Returns:
-        ndarray: The interpolated array
-    """
-
-    #xCoords, yCoords = numpy.meshgrid(yCoords, xCoords)
-
-
-    xIntCoords = xCoords.astype("int")
-    yIntCoords = yCoords.astype("int")
-
-    arrayInt = array[xIntCoords, yIntCoords]
-
-    xGrad = array[
-            (xIntCoords+1).clip(0, array.shape[0]-1), yIntCoords] - arrayInt
-
-    yGrad = array[
-            xIntCoords, (yIntCoords+1).clip(0, array.shape[1]-1)] - arrayInt
-
-    if numpy.any(interpArray):
-        interpArray[:] = (arrayInt + xGrad*(xCoords-xIntCoords)
-                            + yGrad*(yCoords-yIntCoords))
-    else:
-        interpArray = (arrayInt + xGrad*(xCoords-xIntCoords)
-                            + yGrad*(yCoords-yIntCoords))
-
-    return numpy.flipud(numpy.rot90(interpArray.clip(array.min(), array.max())))
-
-def padCropImg(img, newSize):
-    """
-    Pads or Crops a square image to a new square size
-
-    Parameters:
-        img (ndarray): Image to pad/crop
-        newSize (int): New size of square image
-
-    Returns:
-        ndarry: new image
-
-    """
-
-    imgSize = img.shape[0]
-
-    if imgSize==newSize:
-        return img
-
-    coord = int(round(0.5*abs(imgSize-newSize)))
-
-    print(imgSize, newSize)
-    if imgSize>=newSize:
-        newImg = img[coord: -coord, coord: -coord]
-
-    else:
-        newImg = numpy.zeros((newSize, newSize), dtype=img.dtype)
-        newImg[coord: -coord, coord: -coord] = img
-        
-    return newImg
-
-
-#######################
-#WFS Functions
-######################
-
-def findActiveSubaps(subaps, mask, threshold, returnFill=False):
-    '''
-    Finds the subapertures which are "seen" be through the
-    pupil function. Returns the coords of those subaps
-
-    Parameters:
-        subaps (int): The number of subaps in x (assumes square)
-        mask (ndarray): A pupil mask, where is transparent when 1, and opaque when 0
-        threshold (float): The mean value across a subap to make it "active"
-        returnFill (optional, bool): Return an array of fill-factors
-
-    Returns:
-        ndarray: An array of active subap coords
-    '''
-
-    subapCoords = []
-    xSpacing = mask.shape[0]/float(subaps)
-    ySpacing = mask.shape[1]/float(subaps)
-
-    if returnFill:
-        fills = []
-
-    for x in range(subaps):
-        for y in range(subaps):
-            subap = mask[
-                    int(numpy.round(x*xSpacing)): int(numpy.round((x+1)*xSpacing)),
-                    int(numpy.round(y*ySpacing)): int(numpy.round((y+1)*ySpacing))
-                    ]
-
-            if subap.mean() >= threshold:
-                subapCoords.append( [x*xSpacing, y*ySpacing])
-                if returnFill:
-                    fills.append(subap.mean())
-
-    subapCoords = numpy.array( subapCoords )
-
-    if returnFill:
-        return subapCoords, numpy.array(fills)
-    else:
-        return subapCoords
-
-def computeFillFactor(mask, subapPos, subapSpacing):
-
-    fills = numpy.zeros(len(subapPos))
-    for i, (x, y) in enumerate(subapPos):
-        x1 = int(round(x))
-        x2 = int(round(x + subapSpacing))
-        y1 = int(round(y))
-        y2 = int(round(y + subapSpacing))
-        fills[i] = mask[x1:x2, y1:y2].mean()
-
-    return fills
-
-
-def binImgs(data, n):
-    '''
-    Bins one or more images down by the given factor
-    bins. n must be a factor of data.shape, who knows what happens
-    otherwise......
-    '''
-    shape = numpy.array( data.shape )
-
-    n = int(numpy.round(n))
-
-    if len(data.shape)==2:
-        shape[-1]/=n
-        binnedImgTmp = numpy.zeros( shape, dtype=data.dtype )
-        for i in range(n):
-            binnedImgTmp += data[:,i::n]
-        shape[-2]/=n
-        binnedImg = numpy.zeros( shape, dtype=data.dtype )
-        for i in range(n):
-            binnedImg += binnedImgTmp[i::n,:]
-
-        return binnedImg
-    else:
-        shape[-1]/=n
-        binnedImgTmp = numpy.zeros ( shape, dtype=data.dtype )
-        for i in range(n):
-            binnedImgTmp += data[...,i::n]
-
-        shape[-2] /= n
-        binnedImg = numpy.zeros( shape, dtype=data.dtype )
-        for i in range(n):
-            binnedImg += binnedImgTmp[...,i::n,:]
-
-        return binnedImg
-
-def centreOfGravity(img, threshold_frac=0, **kwargs):
-    '''
-    Centroids an image, or an array of images.
-    Centroids over the last 2 dimensions.
-    Sets all values under "threshold_frac*max_value" to zero before centroiding
-    '''
-    if threshold_frac!=0:
-        if len(img.shape)==2:
-            img = numpy.where(img>threshold_frac*img.max(), img, 0 )
-        else:
-            img_temp = (img.T - threshold_frac*img.max(-1).max(-1)).T
-            zero_coords = numpy.where(img_temp<0)
-            img[zero_coords] = 0
-
-    if len(img.shape)==2:
-        y_cent,x_cent = numpy.indices(img.shape)
-        y_centroid = (y_cent*img).sum()/img.sum()
-        x_centroid = (x_cent*img).sum()/img.sum()
-
-    else:
-        y_cent, x_cent = numpy.indices((img.shape[-2],img.shape[-1]))
-        y_centroid = (y_cent*img).sum(-1).sum(-1)/img.sum(-1).sum(-1)
-        x_centroid = (x_cent*img).sum(-1).sum(-1)/img.sum(-1).sum(-1)
-
-    y_centroid+=0.5
-    x_centroid+=0.5
-
-    return numpy.array([y_centroid,x_centroid])
-
-def brtPxlCentroid(img, nPxls, **kwargs):
-    """
-    Centroids using brightest Pixel Algorithm
-    (A. G. Basden et al,  MNRAS, 2011)
-
-    Finds the nPxlsth brightest pixel, subtracts that value from frame,
-    sets anything below 0 to 0, and finally takes centroid.
-
-    Parameters:
-        img (ndarray): 2d or greater rank array of imgs to centroid
-        threshold (float): Percentage of pixels to use for centroid
-    """
-
-    if len(img.shape)==2:
-        pxlValue = numpy.sort(img.flatten())[-nPxls]
-        img-=pxlValue
-        img.clip(0, img.max())
-
-    elif len(img.shape)==3:
-        pxlValues = numpy.sort(
-                        img.reshape(img.shape[0], img.shape[-1]*img.shape[-2])
-                        )[:,-nPxls]
-        img[:]  = (img.T - pxlValues).T
-        img.clip(0, img.max(), out=img)
-
-    return simpleCentroid(img)
-
-def corrConvolve(x, y):
-    '''
-    2D convolution using FFT, use to generate cross-correlations.
-
-    Args:
-        x (array): subap image
-        y (array): reference image
-    Returns:
-        ndarray: cross-correlation of x and y
-    '''
-
-    fr = numpy.fft.fft2(x)
-    fr2 = numpy.fft.fft2(y[::-1, ::-1])
-    m, n = fr.shape
-
-    cc = (numpy.fft.ifft2(fr*fr2)).real
-    cc = numpy.roll(cc, -m/2+1, axis=0)
-    cc = numpy.roll(cc, -n/2+1, axis=1)
-
-    return cc
-
-def correlationCentriod(im, threshold_fac, ref):
-    '''
-    Correlation Centroider, currently only works for 3d im shape.
-    Performs a simple thresholded COM on the correlation.
-
-    Args:
-        im: subap images (t, y, x)
-        threshold_fac: threshold for COM (0=all pixels, 1=brightest pixel)
-        ref: reference image (t, y, x)
-    Returns:
-        ndarray: centroids of im, given as x, y
-    '''
-
-    nt, ny, nx = im.shape
-
-    cents = numpy.zeros((2, nt))
-
-    for frame in range(nt):
-        # Correlate frame with reference image
-        corr = corrConvolve(im[frame], ref[frame])
-
-        # Find brightest pixel.
-        index_y, index_x = numpy.unravel_index(corr.argmax(),
-                                               corr.shape)
-
-        # Apply threshold
-        corr -= corr.min()
-        mx = corr.max()
-        bg = threshold_fac*mx
-        corr = numpy.clip(corr, bg, mx) - bg
-
-        # Centroid
-        s_y, s_x = corr.shape
-        XRAMP = numpy.arange(s_x)
-        YRAMP = numpy.arange(s_y)
-        XRAMP.shape = (1, s_x)
-        YRAMP.shape = (s_y,  1)
-
-        si = corr.sum()
-        if si == 0:
-            si = 1
-        cx = (corr * XRAMP).sum() / si
-        cy = (corr * YRAMP).sum() / si
-
-        cents[:, frame] = cx, cy
-
-    return cents
-
-def quadCell(img, **kwargs):
-
-    xSum = img.sum(-2)
-    ySum = img.sum(-1)
-
-    xCent = xSum[..., 1] - xSum[..., 0]
-    yCent = ySum[..., 1] - ySum[..., 0]
-
-    return numpy.array([xCent, yCent])
-
-def zernike(j, N):
-    """
-     Creates the Zernike polynomial with mode index j,
-     where j = 1 corresponds to piston.
-
-     Args:
-        j (int): The noll j number of the zernike mode
-        N (int): The diameter of the zernike more in pixels
-     Returns:
-        ndarray: The Zernike mode
-     """
-    n, m = zernIndex(j)
-
-    coords = numpy.linspace(-1, 1, N)
-    X, Y = numpy.meshgrid(coords, coords)
-    R = numpy.sqrt(X**2 + Y**2)
-    theta = numpy.arctan2(Y, X)
-
-    if m == 0:
-        Z = numpy.sqrt(n+1)*zernikeRadialFunc(n, 0, R)
-    else:
-        if j%2 == 0: # j is even
-            Z = numpy.sqrt(2*(n+1))*zernikeRadialFunc(n, m, R) * numpy.cos(m*theta)
-        else:   #i is odd
-            Z = numpy.sqrt(2*(n+1))*zernikeRadialFunc(n, m, R) * numpy.sin(m*theta)
-
-
-    return Z * circle(N/2., N)
-
-
-
-def zernikeRadialFunc(n, m, r):
-    """
-    Fucntion to calculate the Zernike radial function
-    """
-
-    R = numpy.zeros(r.shape)
-    for i in xrange(0,int((n-m)/2)+1):
-
-        R += r**(n-2*i) * (((-1)**(i))*numpy.math.factorial(n-i)) / ( numpy.math.factorial(i) * numpy.math.factorial(0.5*(n+m)-i) * numpy.math.factorial(0.5*(n-m)-i) )
-
-    return R
-
-
-
-def zernIndex(j, sign=0):
-    """
-    returns the [n,m] list giving the radial order n and azimutal order
-    of the zernike polynomial of index j
-    if sign is set, will also return a 1 for cos, -1 for sine or 0 when m==0.
-    """
-    n = int((-1.+numpy.sqrt(8*(j-1)+1))/2.)
-    p = (j-(n*(n+1))/2.)
-    k = n%2
-    m = int((p+k)/2.)*2 - k
-    if sign==0:
-        return [n,m]
-    else:#determine whether is sine or cos term.
-        if m!=0:
-            if j%2==0:
-                s=1
-            else:
-                s=-1
-
-        else:
-            s=0
-        return [n,m,s]
-
-
-
-def zernikeArray(J, N):
-    """
-    Creates an array of Zernike Polynomials
-
-    Parameters:
-        maxJ (int or list): Max Zernike polynomial to create, or list of zernikes J indices to create
-        N (int): size of created arrays
-
-    Returns:
-        ndarray: array of Zerkike Polynomials
-    """
-    #If list, make those Zernikes
-    try:
-        nJ = len(J)
-        Zs = numpy.empty((nJ, N, N))
-        for i in xrange(nJ):
-            Zs[i] = zernike(J[i], N)
-
-    #Else, cast to int and create up to that number
-    except TypeError:
-
-        maxJ = int(numpy.round(J))
-        N = int(numpy.round(N))
-
-        Zs = numpy.empty((maxJ, N, N))
-
-        for j in xrange(1,maxJ+1):
-            Zs[j-1] = zernike(j, N)
-
-    return Zs
-
-
-def aziAvg(data):
-    """
-    Measures the azimuthal average of a 2d array
-
-    Parameters:
-
-        data (ndarray): A 2-d array of data
-
-    Returns:
-        ndarray: A 1-d vector of the azimuthal average
-    """
-
-    size = data.shape[0]
-    avg = numpy.empty(size/2, dtype="float")
-    for i in range(size/2):
-        ring = circle(i+1, size) - circle(i, size)
-        avg[i] = (ring*data).sum()/(ring.sum())
-
-    return avg
-
-def calcTiltCorrect(nPxls, wvl, telDiam):
-    """
-    Calculates the required tilt to add to avoid  the PSF being centred
-    on one pixel only
-
-    Parameters:
-        nPxls (int): Number of pixels in array
-        wvl (float): Wavelength of light to correct for in metres
-        telDiam (float): Diameter of telescope in metres
-
-    Returns:
-        ndarray: An array with a tilt to move PSF by 1 pixel
-    """
-
-    # Only required if pxl number is even
-    if not nPxls % 2:
-        # Need to correct for half a pixel angle
-        theta = float(self.FOVrad)/(2*nPxls)
-
-        # Find magnitude of tilt from this angle
-        amplitude = theta * self.telConfig.telDiam/(2*wvl) * 2 * numpy.pi
-
-        coords = numpy.linspace(-1, 1, nPxls)
-        X, Y = numpy.meshgrid(coords, coords)
-        tiltFix = -1*amplitude * (X+Y)
-    else:
-        tiltFix = numpy.zeros(nPxls)
-
-    return tiltFix
-
-
-def photonsPerMag(mag, mask, pxlScale, wvlBand, expTime):
-    '''
-    Calculates the photon flux for a given aperture, star magnitude and wavelength band
-
-    Parameters:
-        mag (float): Star apparent magnitude
-        mask (ndarray): 2-d pupil mask array, 1 is transparent, 0 opaque
-        pxlScale (float): size in metres of each pixel in mask
-        wvlBand (float): length of wavelength band in nanometres
-        expTime (float): Exposure time in seconds
-
-    Returns:
-        float: number of photons
-    '''
-
-    #Area defined in cm, so turn m to cm
-    area = mask.sum() * pxlScale**2 * 100**2
-
-    photonPerSecPerAreaPerWvl = 1000 * (10**(-float(mag)/2.5))
-
-    #Wvl defined in Angstroms
-    photonPerSecPerArea = photonPerSecPerAreaPerWvl * wvlBand*10
-
-    photonPerSec = photonPerSecPerArea * area
-
-    photons = photonPerSec * expTime
-
-    return photons
-
-
-#######################
-#Control Functions
-######################
-
-
-class DelayBuffer(list):
-    '''
-    A delay buffer.
-
-    Each time delay() is called on the buffer, the input value is stored.
-    If the buffer is larger than count, the oldest value is removed and returned.
-    If the buffer is not yet full, a zero of similar shape as the last input
-    is returned.
-    '''
-
-    def delay(self, value, count):
-        self.append(value)
-        if len(self) <= count:
-            result = value*0.0
-        else:
-            for _ in range(len(self)-count):
-                result = self.pop(0)
-        return result
