+++ soapy/aotools/__init__.py
@@ -1,5 +1,7 @@
- 
+
 
 from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
+
+from .aotools import *
