+++ soapy/wfs/base.py
@@ -88,8 +88,8 @@ except ImportError:
     except ImportError:
         raise ImportError("PyAOS requires either pyfits or astropy")
 
-from .. import AOFFT, aoSimLib, LGS, logger, lineofsight
-from ..aotools import centroiders
+from .. import AOFFT, LGS, logger, lineofsight
+from ..aotools import centroiders, circle
 
 # xrange now just "range" in python3.
 # Following code means fastest implementation used in 2 and 3
@@ -134,7 +134,7 @@ class WFS(object):
             self.mask = mask
         # Else we'll just make a circle
         else:
-            self.mask = aoSimLib.circle(
+            self.mask = circle.circle(
                     self.simConfig.pupilSize/2., self.simConfig.simSize,
                     )
 
@@ -168,7 +168,7 @@ class WFS(object):
         if numpy.any(mask):
             self.mask = mask
         else:
-            self.mask = aoSimLib.circle(
+            self.mask = circle.circle(
                     self.simConfig.pupilSize/2., self.simConfig.simSize,
                     )
 
@@ -228,7 +228,7 @@ class WFS(object):
                     )
 
                 # Calculate the zernikes to add
-                self.elongZs = aoSimLib.zernikeArray([2,3,4], self.simConfig.pupilSize)
+                self.elongZs = circle.zernikeArray([2,3,4], self.simConfig.pupilSize)
 
                 # Calculate the radii of the metapupii at for different elong
                 # Layer heights
@@ -312,7 +312,7 @@ class WFS(object):
         pad = ((self.simConfig.simPad,)*2, (self.simConfig.simPad,)*2)
         phaseAddition = numpy.pad(phaseAddition, pad, mode="constant")
 
-        phaseAddition = aoSimLib.zoom(phaseAddition, self.los.nOutPxls)
+        phaseAddition = interp.zoom(phaseAddition, self.los.nOutPxls)
 
         return phaseAddition
 
@@ -421,7 +421,7 @@ class WFS(object):
             # If imat frame, dont want to make it off-axis
             # if iMatFrame:
             #     try:
-            #         iMatPhase = aoSimLib.zoom(scrns, self.los.nOutPxls, order=1)
+            #         iMatPhase = interp.zoom(scrns, self.los.nOutPxls, order=1)
             #         self.los.EField[:] = numpy.exp(1j*iMatPhase*self.los.phs2Rad)
             #     except ValueError:
             #         raise ValueError("If iMat Frame, scrn must be ``simSize``")
